fix: crash when adding pattern for enderio conduit in single player
The conduit stacks use a "conduit" stack component with a reference
to a value in the conduit registry. However, this value in the registry
has a different object instance in server and client.
The pattern cache was shared across client
and server. When trying to serialize the autocraftable
resources in single player, it would try to send the
conduit reference of the client on the server, and it would not be
found in the server registry.
The fix is to make the cache level-dependant.
I hope this doesn't cause a memory leak (weak hash map please)

Fixes #888